+++ src/assets/components/ItemList/ItemList.css
@@ -6,7 +6,6 @@
 
   width: 1200px;
   margin: 0 auto;
-  gap: 24px;
 }
 
 .ItemList-header {
@@ -16,6 +15,7 @@
 
   width: 100%;
   height: 42px;
+  margin-bottom: 24px;
 }
 
 .ItemList-header .utils {
@@ -62,31 +62,5 @@
   justify-content: flex-start;
   gap: 40px 12px;
   width: 100%;
-}
-
-.Pagination {
-  display: flex;
-  gap: 4px;
-}
-
-.Pagination button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  width: 40px;
-  height: 40px;
-  border: 1px solid var(--gray-200);
-  border-radius: 40px;
-  background-color: #ffffff;
-
-  color: var(--gray-500);
-  font-size: 16px;
-  font-weight: 600;
-}
-
-.Pagination button.current {
-  background-color: #2f80ed;
-  border: none;
-  color: var(--gray-100);
+  margin-bottom: 40px;
 }
